fix: move loader to it's own folder
