+++ src/common/widgets/loader/Loader.css
